refactor: Kebab case GoogleAnalytics
