+++ src/event-timer.spec.ts
@@ -1,7 +1,7 @@
 import { EventTimer } from './event-timer';
-import { trackEvent } from './GoogleAnalytics';
+import { trackEvent } from './google-analytics';
 
-jest.mock('./GoogleAnalytics', () => ({
+jest.mock('./google-analytics', () => ({
 	trackEvent: jest.fn(),
 }));
